+++ packages/core/src/config.ts
@@ -1,11 +1,13 @@
 export interface Configurations {
   angleUnit: number
+  defaultColor: string
   publicPath: string
   showWarning: boolean
 }
 
 export default {
   angleUnit: Math.PI,
+  defaultColor: 'blue',
   publicPath: '',
   showWarning: true,
 } as Configurations