+++ packages/demo/barrages/wave_and_particle.js
@@ -3,7 +3,7 @@ module.exports = {
   mounted() {
     this.setInterval(8, (tick) => {
       this.emitBullets(7, (index) => ({
-        display: 'scale?color=red',
+        display: 'scale',
         origin: {
           x: 240,
           y: 280,
@@ -11,12 +11,13 @@ module.exports = {
         },
         state: {
           rho: 32,
-          theta: index / 3.5,
+          color: 'red',
+          face: index / 3.5,
         },
         mutate(tick) {
           this.rho += 2 + tick / 40
-          this.polarLocate(this.rho, this.theta)
-        }
+          this.polarLocate(this.rho, this.face)
+        },
       }))
     })
   }