+++ packages/web-stg/index.d.ts
@@ -1,2 +1,3 @@
 export * from '@stg/core'
 export * from '@stg/utils'
+export * from '@stg/bullets'