+++ packages/web-stg/src/library.js
@@ -1,8 +1,9 @@
-import { Bullet } from '@stg/core'
-import * as builtin from '@stg/bullets'
+import { use } from '@stg/core'
+import bulletPlugin from '@stg/bullets'
 
-Bullet.install(builtin)
+use(bulletPlugin)
 
 export * from '@stg/core'
 export * from '@stg/utils'
+export * from '@stg/bullets'
 export { version } from '../package.json'