+++ packages/docs/guide/CHANGELOG.md
@@ -4,15 +4,33 @@
 
 ### 新特性
 
-- **core:** 新增 `loadImages` 和 `loadAssets` 方法, 用于加载全局资源.
+- **core:** 新增 `loadImages`, `checkImages` 和 `loadAssets` 方法, 用于加载全局资源.
+- **core:** config 新增 `publicPath`, `defaultColor`, `showWarning` 三个属性.
+- **core:** Updater 新增 `$assets`, `$config` 属性, 它们是对应模块的引用.
+- **core:** 子弹模板现在可以使用 Query 字符串, 运行时将使用类似 CSS 的优先级进行渲染.
+- **core:** BulletOptions 中的 `judgeType` 和 `fieldType` 属性都允许传入函数.
+- **core:** 新增了 `point.drawImage` 函数, 用于绘制来自资源文件的图形.
+- **core:** 新增了 `use` 函数, 用于加载插件, 同时 bullets 模块也用插件的方式重新实现.
+- **bullets:** 增加了对东方 Project 第二代引擎 (风神录后) 的部分弹型的支持.
+- **bullets:** 新增了 `colorSeries` 对象, 包含了东方 Project 中默认的颜色系列.
+
+### 重大变化
+
+- **core:** 重写了 Template 模块, 实现了下面的函数:
+  - `defineTemplate`: 加载子弹模板 (相当于过去的 `Bullet.install`).
+  - `buildFromImages`: 从图片资源中直接生成子弹模板.
+- **core:** 新的 BulletTemplate 仅支持 `applied` 和 `display` 两个钩子函数.
+- **core:** BulletOptions 中的 `extends` 属性更名为 `display`, 同时允许传入函数.
+- **bullets:** 将涉及场地和判定相关的函数从 bullets 模块放入 core 模块.
 
 ### 修复与其他
 
-打包时将生成多个入口文件:
-
-- stg.min.css: 所需的样式文件.
-- stg.common.js: 可直接在 NodeJS 环境下加载的模块.
-- stg.runtime.js: 应当在浏览器中加载使用的模块, 内置了所需的样式.
+- 修复了 `$ref` 中默认包含 `base` 而非 `origin` 的问题.
+- 移除了目前项目对 deepmerge 包的依赖.
+- 打包时将生成多个入口文件:
+  - stg.min.css: 所需的样式文件.
+  - stg.common.js: 可直接在 NodeJS 环境下加载的模块.
+  - stg.runtime.js: 应当在浏览器中加载使用的模块, 内置了所需的样式.
 
 ## v0.3
 
@@ -26,7 +44,7 @@
 
 - **core:** Looping 构造函数新增 `skipFrame` 参数, 可用于调整描画间隔.
 - **core:** 新增 `looping.initialize()` 方法, 可以用于重新初始化实例.
-- **core:** 新增 `core.config.angleUnit`, 用于配置采用的角度单位, 默认为 `Math.PI`.
+- **core:** 新增 `config.angleUnit`, 用于配置采用的角度单位, 默认为 `Math.PI`.
 - **core:** `looping.getStatus()` 返回对象新增 `frameRate` 属性, 表示当前物理帧数.
 
 ## v0.2.1